+++ src/main/java/org/example/myfinalproject/CommissionSalariedEmployee.java
@@ -5,15 +5,21 @@ import java.time.LocalDate;
 
 public class CommissionSalariedEmployee extends Employee implements Serializable, CalculateSalary, MonthlyEarning {
     private double commissionRate;
-    private double numberOfSales;
+    private int numberOfSales;
 
-    public CommissionSalariedEmployee(String firstName, String lastName, String nationalID, String phoneNumber, double managerBonus, LocalDate startDate, LocalDate birthdate, Department department, double commissionRate, double numberOfSales) {
+    public CommissionSalariedEmployee(String firstName, String lastName, String nationalID,
+                                      String phoneNumber, double managerBonus, LocalDate startDate,
+                                      LocalDate birthdate, Department department, double commissionRate, int numberOfSales) {
+
         super(firstName, lastName, nationalID, phoneNumber, managerBonus, startDate, birthdate, department);
         this.commissionRate = commissionRate;
         this.numberOfSales = numberOfSales;
     }
 
-    public CommissionSalariedEmployee(String firstName, String lastName, String nationalID, String phoneNumber, LocalDate startDate, LocalDate birthdate, Department department, double commissionRate, double numberOfSales) {
+    public CommissionSalariedEmployee(String firstName, String lastName, String nationalID,
+                                      String phoneNumber, LocalDate startDate, LocalDate birthdate,
+                                      Department department, double commissionRate, int numberOfSales) {
+
         super(firstName, lastName, nationalID, phoneNumber, startDate, birthdate, department);
         this.commissionRate = commissionRate;
         this.numberOfSales = numberOfSales;
@@ -28,11 +34,11 @@ public class CommissionSalariedEmployee extends Employee implements Serializable
         this.commissionRate = commissionRate;
     }
 
-    public double getNumberOfSales() {
+    public int getNumberOfSales() {
         return numberOfSales;
     }
 
-    public void setNumberOfSales(double numberOfSales) {
+    public void setNumberOfSales(int numberOfSales) {
         this.numberOfSales = numberOfSales;
     }
